数据类型、引入编码

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesProductLedger.java
@@ -38,7 +38,7 @@ public class IcesProductLedger extends BaseEntity {
     /**
      * 库存量
      */
-    private Long plStock;
+    private Float plStock;
     /**
      * 已删除
      */
